LibSerenity: Use DRM API to setup two GPU buffers on OpenBSD
When using inteldrm, we can't do a simple mmap() on the device like
we can with efifb.  Implement enough of the DRM API to allocate
two buffers on the GPU that we can now use for double buffering.

The code for doing a single buffer mmap should probably be kept as
a fallback in case there is no accelerated graphics on the machine.